Changed time endiannes to little endian. Fix linux build
