+++ unit_tests/mission/Experiments/payload/PayloadExperimentTest.cpp
@@ -1,6 +1,6 @@
 #include "gtest/gtest.h"
 #include "gmock/gmock.h"
-#include "OSMock.hpp"
+#include "OsMock.hpp"
 #include "eps/hk.hpp"
 #include "experiment/payload/payload_exp.hpp"
 #include "mock/FsMock.hpp"
